# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AvailabilityBlockRequest": {
        "properties": {
          "ends_at": {
            "format": "date-time",
            "title": "Ends At",
            "type": "string"
          },
          "quantity": {
            "exclusiveMinimum": 0.0,
            "title": "Quantity",
            "type": "integer"
          },
          "reason": {
            "maxLength": 120,
            "minLength": 2,
            "title": "Reason",
            "type": "string"
          },
          "starts_at": {
            "format": "date-time",
            "title": "Starts At",
            "type": "string"
          }
        },
        "required": [
          "starts_at",
          "ends_at",
          "quantity",
          "reason"
        ],
        "title": "AvailabilityBlockRequest",
        "type": "object"
      },
...
      "CategoryResponse": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "required_attributes": {
            "additionalProperties": true,
            "title": "Required Attributes",
            "type": "object"
          },
          "risk_tier": {
            "title": "Risk Tier",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "risk_tier",
          "required_attributes"
        ],
        "title": "CategoryResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ListingCreateRequest": {
        "properties": {
          "attributes": {
            "additionalProperties": true,
            "title": "Attributes",
            "type": "object"
          },
          "category_id": {
            "format": "uuid",
            "title": "Category Id",
            "type": "string"
          },
          "condition": {
            "enum": [
              "new",
              "like_new",
              "good",
              "fair"
            ],
            "title": "Condition",
            "type": "string"
          },
          "delivery_enabled": {
            "default": false,
            "title": "Delivery Enabled",
            "type": "boolean"
          },
          "description": {
            "maxLength": 5000,
            "minLength": 10,
            "title": "Description",
            "type": "string"
          },
          "latitude": {
            "anyOf": [
              {
                "maximum": 90.0,
                "minimum": -90.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": -180.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "pickup_enabled": {
            "default": true,
            "title": "Pickup Enabled",
            "type": "boolean"
          },
          "prices": {
            "items": {
              "$ref": "#/components/schemas/PriceInput"
            },
            "maxItems": 4,
            "minItems": 1,
            "title": "Prices",
            "type": "array"
          },
          "public_locality": {
            "maxLength": 160,
            "minLength": 2,
            "title": "Public Locality",
            "type": "string"
          },
          "quantity": {
            "default": 1,
            "exclusiveMinimum": 0.0,
            "maximum": 1000.0,
            "title": "Quantity",
            "type": "integer"
          },
          "title": {
            "maxLength": 160,
            "minLength": 3,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "category_id",
          "title",
          "description",
          "condition",
          "public_locality",
          "prices"
        ],
        "title": "ListingCreateRequest",
        "type": "object"
      },
      "ListingResponse": {
        "properties": {
          "attributes": {
            "additionalProperties": true,
            "title": "Attributes",
            "type": "object"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryResponse"
          },
          "condition": {
            "title": "Condition",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "delivery_enabled": {
            "title": "Delivery Enabled",
            "type": "boolean"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "owner_user_id": {
            "format": "uuid",
            "title": "Owner User Id",
            "type": "string"
          },
          "pickup_enabled": {
            "title": "Pickup Enabled",
            "type": "boolean"
          },
          "prices": {
            "items": {
              "$ref": "#/components/schemas/PriceResponse"
            },
            "title": "Prices",
            "type": "array"
          },
          "public_locality": {
            "title": "Public Locality",
            "type": "string"
          },
          "quantity": {
            "title": "Quantity",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "owner_user_id",
          "category",
          "title",
          "description",
          "condition",
          "quantity",
          "attributes",
          "status",
          "pickup_enabled",
          "delivery_enabled",
          "public_locality",
          "version",
          "prices",
          "created_at"
        ],
        "title": "ListingResponse",
        "type": "object"
      },
      "ListingUpdateRequest": {
        "properties": {
          "condition": {
            "anyOf": [
              {
                "enum": [
                  "new",
                  "like_new",
                  "good",
                  "fair"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Condition"
          },
          "delivery_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivery Enabled"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 5000,
                "minLength": 10,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "pickup_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pickup Enabled"
          },
          "public_locality": {
            "anyOf": [
              {
                "maxLength": 160,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Public Locality"
          },
          "quantity": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 1000.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 160,
                "minLength": 3,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "version": {
            "exclusiveMinimum": 0.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version"
        ],
        "title": "ListingUpdateRequest",
        "type": "object"
      },
...
      "SellerHeaderRequest": {
        "properties": {
          "description": {
            "default": "",
            "maxLength": 1200,
            "title": "Description",
            "type": "string"
          },
          "display_name": {
            "maxLength": 120,
            "minLength": 2,
            "title": "Display Name",
            "type": "string"
          },
          "operating_hours": {
            "additionalProperties": true,
            "title": "Operating Hours",
            "type": "object"
          },
          "public_locality": {
            "maxLength": 160,
            "minLength": 2,
            "title": "Public Locality",
            "type": "string"
          }
        },
        "required": [
          "display_name",
          "public_locality"
        ],
        "title": "SellerHeaderRequest",
        "type": "object"
      },
      "SellerHeaderResponse": {
        "properties": {
          "description": {
            "default": "",
            "maxLength": 1200,
            "title": "Description",
            "type": "string"
          },
          "display_name": {
            "maxLength": 120,
            "minLength": 2,
            "title": "Display Name",
            "type": "string"
          },
          "moderation_status": {
            "title": "Moderation Status",
            "type": "string"
          },
          "operating_hours": {
            "additionalProperties": true,
            "title": "Operating Hours",
            "type": "object"
          },
          "public_locality": {
            "maxLength": 160,
            "minLength": 2,
            "title": "Public Locality",
            "type": "string"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "display_name",
          "public_locality",
          "user_id",
          "moderation_status"
        ],
        "title": "SellerHeaderResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "4by4 For Hire API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/listings": {
      "get": {
        "operationId": "search_listings_api_v1_listings_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "category_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Id"
            }
          },
          {
            "in": "query",
            "name": "latitude",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 90,
                  "minimum": -90,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Latitude"
            }
          },
          {
            "in": "query",
            "name": "longitude",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maximum": 180,
                  "minimum": -180,
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Longitude"
            }
          },
          {
            "in": "query",
            "name": "radius_km",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 200,
              "minimum": 1,
              "title": "Radius Km",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ListingResponse"
                  },
                  "title": "Response Search Listings Api V1 Listings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Listings",
        "tags": [
          "catalog"
        ]
      },
      "post": {
        "operationId": "create_listing_api_v1_listings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListingCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Listing",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/listings/{listing_id}": {
      "get": {
        "operationId": "get_listing_api_v1_listings__listing_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Listing Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Listing",
        "tags": [
          "catalog"
        ]
      },
      "patch": {
        "operationId": "update_listing_api_v1_listings__listing_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Listing Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListingUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Listing",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/listings/{listing_id}/archive": {
      "post": {
        "operationId": "archive_listing_api_v1_listings__listing_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Listing Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Archive Listing",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/listings/{listing_id}/availability-blocks": {
      "post": {
        "operationId": "add_availability_block_api_v1_listings__listing_id__availability_blocks_post",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Listing Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvailabilityBlockRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Add Availability Block Api V1 Listings  Listing Id  Availability Blocks Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Add Availability Block",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/listings/{listing_id}/pause": {
      "post": {
        "operationId": "pause_listing_api_v1_listings__listing_id__pause_post",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Listing Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Pause Listing",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/listings/{listing_id}/publish": {
      "post": {
        "operationId": "publish_listing_api_v1_listings__listing_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Listing Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Publish Listing",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/api/v1/listings/{listing_id}/resume": {
      "post": {
        "operationId": "resume_listing_api_v1_listings__listing_id__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "listing_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Listing Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Resume Listing",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/me/listings": {
      "get": {
        "operationId": "my_listings_api_v1_me_listings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ListingResponse"
                  },
                  "title": "Response My Listings Api V1 Me Listings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "My Listings",
        "tags": [
          "catalog"
        ]
      }
    },
    "/api/v1/me/seller-header": {
      "put": {
        "operationId": "upsert_seller_header_api_v1_me_seller_header_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SellerHeaderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SellerHeaderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upsert Seller Header",
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/api/v1/sellers/{user_id}": {
      "get": {
        "operationId": "public_seller_api_v1_sellers__user_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SellerHeaderResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Public Seller",
        "tags": [
          "catalog"
        ]
      }
    },
...
  }
}